requirements.txt
Jinja2==3.1.5
MarkupSafe==3.0.2
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
geocoder==1.38.1
idna==3.10
jedi==0.19.2
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-leetcode==1.2.1
python-lsp-server[all]==1.12.0
pywal==3.3.0
pywalfox==2.7.4
requests==2.32.3
requests==2.32.3
s-tui==1.1.6
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
